backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
alembic==1.18.4
python-multipart==0.0.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==5.0.0
httpx==0.27.2